+++ src/domain/transaction/transaction.mts
@@ -295,56 +295,3 @@ export class Transaction {
     return tx
   }
 }
-
-class UnspentTransactionOut {
-  private _blockhash: string
-  private _txid: string
-  private _index: number
-  private _output: Output
-
-  public get blockhash() {
-    return this._blockhash
-  }
-
-  public get txid() {
-    return this._txid
-  }
-
-  public get index() {
-    return this._index
-  }
-
-  public get output() {
-    return this._output
-  }
-
-  constructor({ blockhash, txid, index, output }: { blockhash: string, txid: string; index: number; output: Output }) {
-    this._blockhash = blockhash
-    this._txid = txid
-    this._index = index
-    this._output = output
-  }
-
-  public display(): {
-    blockhash: string
-    txid: string
-    index: number
-    amount: number
-    publicKey: string
-  } {
-    return {
-      blockhash: this._blockhash,
-      txid: this._txid,
-      index: this._index,
-      amount: this._output.amount,
-      publicKey: hex(this._output.publicKey)
-    }
-  }
-
-  public static fromTransaction(blockhash: Uint8Array, tx: Transaction): UnspentTransactionOut[] {
-    return tx.outputs.map((output, index) => new UTxOut({ blockhash: hex(blockhash), txid: hex(tx.id), index, output }))
-  }
-}
-
-export const UTxOut = UnspentTransactionOut
-export type UTxOut = UnspentTransactionOut
