--- src/lib/p2p/index.mts
@@ -1,3 +1,0 @@
-export { Server } from './server.mts'
-export { Peer as Client } from './peer.mts'
-export type { Session } from './peer.mts'
